[migration] Add a separate container for migration
