+++ migration/entrypoint.sh
@@ -1,0 +1,2 @@
+prefect server start --detach
+#systemctl start datapusher.service
