+++ migration/requirements.txt
@@ -1,0 +1,13 @@
+prefect==2.14.15
+pathlib==1.0.1
+semver==3.0.2
+pydantic==2.5.3
+datasize==1.0.0
+python-dateutil==2.8.2
+pytz==2023.3.post1
+python-dotenv==1.0.0
+psycopg2-binary==2.9.9
+boto3==1.34.35
+pandas==2.2.0
+openpyxl==3.1.2
+lxml==5.1.0
