+++ migration/tasks/migration_task.py
@@ -1,0 +1,6 @@
+from prefect import flow,task
+
+@task
+def migration_task():
+    msg = 'Hello from Task'
+    return msg
